@@ -1,5 +1,6 @@
 // app/layout.tsx
 import './globals.css';
+import Link from 'next/link';
 import { getSiteName, getSiteUrl } from '@/lib/site';
 import LayoutWrapper from '@/components/LayoutWrapper';
 import { getServerSupabase } from '@/lib/supabase/server';
@@ -39,7 +40,7 @@ export default async function RootLayout({
       <body className="min-h-screen bg-[var(--bg)] text-[var(--fg)]">
         <header className="sticky top-0 z-40 bg-[var(--card)] border-b-2 border-[var(--border)] h-16">
           <div className="max-w-6xl mx-auto px-4 h-full flex items-center justify-between">
-            <a href="/" className="text-heading text-[var(--fg)]">Enlist</a>
+            <Link href="/" className="text-heading text-[var(--fg)]">Enlist</Link>
             <HeaderNavigation isAuthenticated={Boolean(session)} />
           </div>
         </header>
