@@ -1,7 +1,28 @@
 // lib/user-profile.ts
 import { createClient } from '@supabase/supabase-js';
 import { readFileSync } from 'fs';
-import path from 'path';
+
+interface GriffinEmploymentHistoryEntry {
+  job_title?: string;
+  employer_name?: string;
+  start_date?: string;
+  end_date?: string;
+}
+
+interface GriffinProfile {
+  personal_information?: {
+    full_name?: string;
+    email?: string;
+    phone?: string;
+    address?: string;
+  };
+  skills_and_qualifications?: {
+    technical_skills?: string[];
+    business_and_product_skills?: string[];
+  };
+  employment_history?: GriffinEmploymentHistoryEntry[];
+  resume_location?: string;
+}
 
 function getSupabaseAdmin() {
   const supabaseUrl = process.env.NEXT_PUBLIC_SUPABASE_URL!;
@@ -66,7 +87,7 @@ export async function importProfileFromFile(userId: string, filePath?: string): 
     // Default to the griffin.json file if no path provided
     const profilePath = filePath || '/Users/griffinlong/Projects/personal_projects/resume/griffin.json';
     
-    const profileData = JSON.parse(readFileSync(profilePath, 'utf8'));
+    const profileData = JSON.parse(readFileSync(profilePath, 'utf8')) as GriffinProfile;
     
     // Transform griffin.json structure to our UserProfile interface
     const userProfile: UserProfile = {
@@ -82,11 +103,11 @@ export async function importProfileFromFile(userId: string, filePath?: string): 
           ...(profileData.skills_and_qualifications?.technical_skills || []),
           ...(profileData.skills_and_qualifications?.business_and_product_skills || [])
         ],
-        years_experience: calculateYearsExperience(profileData.employment_history || []),
-        previous_roles: (profileData.employment_history || []).map((job: any) => ({
-          title: job.job_title || '',
-          company: job.employer_name || '',
-          duration: `${job.start_date || ''} - ${job.end_date || 'Present'}`
+        years_experience: calculateYearsExperience(profileData.employment_history ?? []),
+        previous_roles: (profileData.employment_history ?? []).map((job) => ({
+          title: job.job_title ?? '',
+          company: job.employer_name ?? '',
+          duration: formatEmploymentDuration(job.start_date, job.end_date)
         }))
       },
       preferences: {
@@ -176,21 +197,44 @@ export async function getOrCreateUserProfile(userId: string): Promise<UserProfil
 }
 
 // Calculate years of experience from employment history
-function calculateYearsExperience(employmentHistory: any[]): number {
+function calculateYearsExperience(employmentHistory: GriffinEmploymentHistoryEntry[]): number {
   if (!employmentHistory || employmentHistory.length === 0) return 0;
-  
+
   let totalMonths = 0;
-  
+
   for (const job of employmentHistory) {
-    if (job.start_date && job.end_date) {
-      const start = new Date(job.start_date);
-      const end = job.end_date === 'Present' ? new Date() : new Date(job.end_date);
-      const months = (end.getFullYear() - start.getFullYear()) * 12 + (end.getMonth() - start.getMonth());
-      totalMonths += months;
+    const { start_date: startDate, end_date: endDate } = job;
+    if (!startDate) {
+      continue;
     }
+
+    const start = new Date(startDate);
+    const normalizedEndDate = endDate?.trim();
+    const end = !normalizedEndDate || normalizedEndDate.toLowerCase() === 'present'
+      ? new Date()
+      : new Date(normalizedEndDate);
+    if (Number.isNaN(start.getTime()) || Number.isNaN(end.getTime())) {
+      continue;
+    }
+
+    const months = (end.getFullYear() - start.getFullYear()) * 12 + (end.getMonth() - start.getMonth());
+    totalMonths += Math.max(months, 0);
   }
-  
+
   return Math.round(totalMonths / 12);
+}
+
+function formatEmploymentDuration(startDate?: string, endDate?: string): string {
+  if (!startDate && !endDate) {
+    return '';
+  }
+
+  if (!startDate) {
+    return endDate ?? 'Present';
+  }
+
+  const formattedEnd = endDate ?? 'Present';
+  return `${startDate} - ${formattedEnd}`;
 }
 
 // Validate user profile completeness
